* Applied 64 bit fix by Goetz

+++ plugins/rtjpeg/mmx.h
@@ -44,8 +44,8 @@
 	 they be truncated by the compiler)
 */
 typedef	union {
-	long long		q;	/* Quadword (64-bit) value */
-	unsigned long long	uq;	/* Unsigned Quadword */
+	int64_t			q;	/* Quadword (64-bit) value */
+	uint64_t		uq;	/* Unsigned Quadword */
 	int			d[2];	/* 2 Doubleword (32-bit) values */
 	unsigned int		ud[2];	/* 2 Unsigned Doubleword */
 	short			w[4];	/* 4 Word (16-bit) values */
